Full Redesign
I completely redesigned the website. I didn't add any new html files but I completely changed the HTML in those files in order for me to do the redesign.

For the redesign, I tweaked the layout of the pages as well as the colour scheme and aesthetic. I wanted my website to have a low-key Genshin Impact theme to it.