Side navigation text appears upon hovering icon
I added some sloppily put together HTML, CSS and JavaScript so that when the user hovers over the icons on the side navigation bar, text pops out from the right of the sidebar displaying the name of the page the link leads to.

I'm planning on adding a slide effect to make it look cooler.

Issues:
For some reason, it works on every page except the home page.