Redesigned Subpages
I redesigned my profile and projects page.

For the  profile page, rather than having all the content on top of one another on desktop resolution, the content will be split into 2 parts, LHS is the profile picture, basic info such as name, birthday & occupation. RHS is the more descriptive side with a list of my favourite games, music artists and anime.

I changed the projects page to have a 3-column grid to display the items. Still a WIP.