Re: Contacts Page
The sidebar/side nav was bugged so I just remade the contacts page from scratch while trying to achieve the same appearance.

One difference is that the hover now also changes the colour of the PNG in the contact-item-list. Also I added some JS so when the user clicks on one of the items, a window popup will appear to confirm if they wish to visit another website.